requirements.txt
# required packages
python-dotenv
databricks-sql-connector
databricks-sdk
websockets
fastapi
uvicorn
web3
